WIP -- restructure and list tests needed
