+++ modules/contracts/src.ts/tests/assetTransfer.spec.ts
@@ -1,0 +1,6 @@
+describe("AssetTransfer.sol", () => {
+  it.skip("should successfully transfer eth", async () => {});
+  it.skip("should successfully transfer erc20", async () => {});
+  it.skip("should successfully transfer assets if `transfer` does not return boolean", async () => {});
+  it.skip("should make funds withdrawable if transfer fails", async () => {});
+});
