+++ modules/contracts/src.ts/tests/channelFactory.spec.ts
@@ -1,12 +1,13 @@
 import { getCreate2MultisigAddress, getPublicIdentifierFromPublicKey, expect } from "@connext/vector-utils";
 import { AddressZero } from "@ethersproject/constants";
-import { Contract, ContractFactory, BigNumber } from "ethers";
+import { Contract, BigNumber } from "ethers";
 import pino from "pino";
 
-import { ChannelMastercopy, ChannelFactory, Proxy } from "../artifacts";
+import { ChannelMastercopy, Proxy } from "../artifacts";
 import { VectorChainReader } from "../services";
 
 import { alice, bob, provider } from "./constants";
+import { createTestChannel, createTestChannelFactory } from "./utils";
 
 describe("ChannelFactory", () => {
   const alicePubId = getPublicIdentifierFromPublicKey(alice.publicKey);
@@ -19,21 +20,13 @@ describe("ChannelFactory", () => {
   beforeEach(async () => {
     chainId = (await provider.getNetwork()).chainId;
 
-    channelMastercopy = await new ContractFactory(ChannelMastercopy.abi, ChannelMastercopy.bytecode, alice).deploy();
-    await channelMastercopy.deployed();
-
-    channelFactory = await new ContractFactory(ChannelFactory.abi, ChannelFactory.bytecode, alice).deploy(
-      channelMastercopy.address,
-    );
-    await channelFactory.deployed();
+    const deployRes = await createTestChannelFactory();
+    channelMastercopy = deployRes.channelMastercopy;
+    channelFactory = deployRes.channelFactory;
 
     const network = await provider.getNetwork();
     const chainProviders = { [network.chainId]: provider };
-    chainReader = new VectorChainReader(
-      chainProviders,
-      pino().child({ module: "VectorChainReader" }),
-    );
-
+    chainReader = new VectorChainReader(chainProviders, pino().child({ module: "VectorChainReader" }));
   });
 
   it("should deploy", async () => {
@@ -48,14 +41,17 @@ describe("ChannelFactory", () => {
     expect(await channelFactory.proxyCreationCode()).to.equal(Proxy.bytecode);
   });
 
+  it.skip("should return the correctly calculated channel address", async () => {});
+
+  it("should get the participants from a deployed channel", async () => {
+    const channel = await createTestChannel(channelFactory);
+    const participants = await channel.getParticipants();
+    expect(participants[0]).to.equal(alice.address);
+    expect(participants[1]).to.equal(bob.address);
+  });
+
   it("should create a channel", async () => {
-    const created = new Promise(res => {
-      channelFactory.once(channelFactory.filters.ChannelCreation(), res);
-    });
-    const tx = await channelFactory.createChannel(alice.address, bob.address, chainId);
-    expect(tx.hash).to.be.a("string");
-    await tx.wait();
-    const channelAddress = await created;
+    const channel = await createTestChannel(channelFactory);
     const computedAddr = await getCreate2MultisigAddress(
       alicePubId,
       bobPubId,
@@ -63,8 +59,7 @@ describe("ChannelFactory", () => {
       channelFactory.address,
       chainReader,
     );
-    expect(channelAddress).to.be.a("string");
-    expect(channelAddress).to.be.eq(computedAddr.getValue());
+    expect(channel.address).to.be.eq(computedAddr.getValue());
   });
 
   it("should create a channel with a deposit", async () => {
