+++ modules/contracts/src.ts/tests/cmcs/accountant.spec.ts
@@ -1,0 +1,3 @@
+describe("CMCAccountant.sol", () => {
+  it.skip("should properly transfer the balance", async () => {});
+});
